+++ src/App.js
@@ -1,6 +1,4 @@
 import './index.scss';
-import { ModalWindow } from './projects/Modal/ModalWindow';
-import { Counter } from './projects/Counter';
 
 function App() {
 	return <div className='App'></div>;
